+++ doc/DocApp.js
@@ -2,11 +2,13 @@ import React from 'react';
 import 'src/styles/index.scss';
 
 import VisualElements from './VisualElements';
+import RowComponents from './RowComponents';
 
 function DocApp() {
     return (
         <div>
             <VisualElements />
+            <RowComponents />
         </div>
     );
 }